move staking hooks template to types
